+++ src/predictor/localsum.c
@@ -6,7 +6,7 @@
 	Chapter 4.4 in CCSDS 123 Issue 2
 */
 
-void wideNeighborLocalSum(unsigned long * sample, long * localsum, int x, int y, int z, struct arguments * parameters) {
+void wideNeighborLocalSum(uint32_t * sample, int32_t * localsum, uint16_t x, uint16_t y, uint16_t z, struct arguments * parameters) {
 	if (x+y != 0) {
 		if (y == 0 && x > 0){
 			localsum[offset(x,y,z,parameters)] = 4 * sample[offset(x-1,y,z, parameters)];
@@ -20,7 +20,7 @@ void wideNeighborLocalSum(unsigned long * sample, long * localsum, int x, int y,
 	}
 }
 
-void narrowNeighborLocalSum(unsigned long * sample, long * localsum, int smid, int x, int y, int z, struct arguments * parameters) {
+void narrowNeighborLocalSum(uint32_t * sample, int32_t * localsum, int smid, uint16_t x, uint16_t y, uint16_t z, struct arguments * parameters) {
 	if(x+y != 0) {
 		if (y > 0 && x > 0 && x < parameters->xSize-1) {
 			localsum[offset(x,y,z,parameters)] = sample[offset(x-1,y-1,z, parameters)] +  (2 * sample[offset(x,y-1,z, parameters)]) + sample[offset(x+1,y-1,z, parameters)];
@@ -36,7 +36,7 @@ void narrowNeighborLocalSum(unsigned long * sample, long * localsum, int smid, i
 	}
 }
 
-void wideColumnLocalSum(unsigned long * sample, long * localsum, int x, int y, int z, struct arguments * parameters) {
+void wideColumnLocalSum(uint32_t * sample, int32_t * localsum, uint16_t x, uint16_t y, uint16_t z, struct arguments * parameters) {
 	if(!(x == 0 && y == 0)) {
 		if(y > 0) {
 			localsum[offset(x,y,z,parameters)] = 4 * sample[offset(x,y-1,z, parameters)];
@@ -48,7 +48,7 @@ void wideColumnLocalSum(unsigned long * sample, long * localsum, int x, int y, i
 	}
 }
 
-void narrowColumnLocalSum(unsigned long * sample, long * localsum, int smid, int x, int y, int z, struct arguments * parameters) {
+void narrowColumnLocalSum(uint32_t * sample, int32_t * localsum, int smid, uint16_t x, uint16_t y, uint16_t z, struct arguments * parameters) {
 	if(!(x == 0 && y == 0)) {
 		if (y > 0) {
 			localsum[offset(x,y,z,parameters)] = 4 * sample[offset(x,y-1,z, parameters)];
